requirements.txt
Pillow
tkinterdnd2
pypdf
docx2pdf